move datepicker and interval dropdown to client/components
