+++ client/components/stats-interval-dropdown/style.scss
